requirements.txt
chromadb
langchain
flask
GitPython
python-dotenv
sentence-transformers
langchain-community
langchain-huggingface

